tests, net: Move 'mac' helpers to `libnet`
