+++ tests/libnet/mac.go
@@ -1,4 +1,4 @@
-package network
+package libnet
 
 import (
 	cryptorand "crypto/rand"
@@ -7,7 +7,8 @@ import (
 
 func GenerateRandomMac() (net.HardwareAddr, error) {
 	prefix := net.HardwareAddr{0x02, 0x00, 0x00} // local unicast prefix
-	suffix := make(net.HardwareAddr, 3)
+	const macByteSize = 3
+	suffix := make(net.HardwareAddr, macByteSize)
 	_, err := cryptorand.Read(suffix)
 	if err != nil {
 		return nil, err
